Move margin and padding styles from slots to all drawables
