--- lacci/lib/shoes/spacing.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-class Shoes
-  module Spacing
-    def self.included(includer)
-      includer.shoes_styles :margin, :padding, :margin_top, :margin_left, :margin_right, :margin_bottom, :options
-    end
-  end
-end
